Moved Telop Pose Selector location
