+++ SeasonCode2025/src/main/java/frc/robot/RobotContainer.java
@@ -32,7 +32,6 @@ import frc.robot.CatzSubsystems.CatzLEDs.CatzLED;
 import frc.robot.CatzSubsystems.CatzOuttake.CatzOuttake;
 import frc.robot.Commands.DriveAndRobotOrientationCmds.TeleopDriveCmd;
 import frc.robot.Utilities.Alert;
-import frc.robot.Utilities.TeleopPosSelector;
 import frc.robot.Utilities.Alert.AlertType;
 import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj2.command.button.Trigger;
