+++ SeasonCode2025/src/main/java/frc/robot/TeleopPosSelector.java
@@ -5,7 +5,7 @@
 // license that can be found in the LICENSE file at
 // the root directory of this project.
 
-package frc.robot.Utilities;
+package frc.robot;
 
 import java.util.function.Supplier;
 
@@ -22,8 +22,8 @@ import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
 import frc.robot.FieldConstants.Reef;
+import frc.robot.Utilities.AllianceFlipUtil;
 import frc.robot.CatzSubsystems.CatzSuperstructure.LeftRight;
-import frc.robot.RobotContainer;
 import frc.robot.CatzSubsystems.CatzDriveAndRobotOrientation.CatzRobotTracker;
 import frc.robot.CatzSubsystems.CatzDriveAndRobotOrientation.Drivetrain.DriveConstants;
 import frc.robot.Commands.DriveAndRobotOrientationCmds.TrajectoryDriveCmd;
